requirements.txt
aioprocessing==2.0.1
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary>=2.9
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1 
setproctitle==1.3.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
